+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Changes tracked until 2022-07-06.
 ### Unsorted
 
 - data, keepData, dataMap
+- In form, .up-active is given to the button or field that triggered the submission/validation, not the form